Strict font names rule (#25)
Adds strict font rules. This allows for a configuration like this:

```json
{
  "paths": {
    "Modules": {
      "rules": [
        "strict_font_names"
      ],
      "rules_config": {
        "strict_font_names": {
          "allowed_fonts": [
            "LyftProUI-Bold",
            "LyftProUI-BoldItalic",
            "ProximaNova-Regular",
            "ProximaNova-RegularItalic",
            "ProximaNova-Medium",
            "ProximaNova-MediumItalic",
            "ProximaNova-Bold",
            "ProximaNova-BoldItalic"
          ],
          "allow_system_fonts": false
        }
      }
    }
  }
}
```

It simply compares the name of `<font>` and `<fontDescription>`’s `name` attribute to the allowed list. If the node is missing `name` it is a system font. `<font>` and `<fontDescription>` represent system fonts different, but that doesn't matter for our purposes.