Remove tag_name redefinition
This was included by mistake and is unnecessary for current uses as the element's tag will the equivalent to the original tag_name that was used for the search. However, by removing it we can now use nested types.

I added an example of enforcing properties for a stackView's color. This triggers on the following example:
```
<stackView opaque="NO" contentMode="scaleToFill" id="zvX-8E-dZy">
    <color key="backgroundColor" name="+ Background Primary"/>
    [...]
</stackView>
```