feature: specify the color property (e.g. textColor, tintColor) that is causing the named_colors warning. (#65)
Added the specific color property that is causing the warning in the error message.

```
Use of custom textColors is not allowed. Use a named color instead. [rule: named_colors]
Use of custom tintColors is not allowed. Use a named color instead. [rule: named_colors]
Use of custom backgroundColors is not allowed. Use a named color instead. [rule: named_colors]

"foo" is not one of the allowed backgroundColors: `bar`,`baz`. [rule: named_colors]
```

This way when the developer uses the `object_id` to find the problematic object, they don't need to look at the xml to know which color property (e.g backgroundColor, textColor, tintColor) to modify.